requirements.txt
pyrogram
python-dotenv
aiofiles
tgcrypto
Pillow
meval